\name{add.grid}
\alias{add.grid}
\title{Add grid to x axis}
\usage{
  add.grid(side = c("top", "bottom", "left", "right"), ...,
    ticks = c("default", "yes", "no"), scales, components,
    line.col, where = "both")
}
\arguments{
  \item{side}{}

  \item{...}{}

  \item{ticks}{}

  \item{scales}{}

  \item{components}{}

  \item{line.col}{}
}
\description{
  Add grid to x axis
}
\references{
  Forked from \code{latticeExtra::axis.grid()}.
}
\keyword{internal}

