% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rp_helpers.R
\name{as.character.rapport.meta}
\alias{as.character.rapport.meta}
\title{Convert Metadata to Character}
\usage{
\method{as.character}{rapport.meta}(x, ...)
}
\arguments{
\item{x}{template metadata object}

\item{...}{ignored}
}
\description{
Converts template metadata to character vector with YAML strings.
}

