\name{as.character.rp.inputs}
\alias{as.character.rp.inputs}
\title{Convert Inputs to Character}
\usage{
  \method{as.character}{rp.inputs} (x, ...)
}
\arguments{
  \item{x}{template inputs object}

  \item{...}{ignored}
}
\description{
  Converts template inputs to character vector with YAML
  strings.
}

