\name{as.character.rp.meta}
\alias{as.character.rp.meta}
\title{Convert Metadata to Character}
\usage{
  as.character.rp.meta(meta, include.examples = TRUE)
}
\arguments{
  \item{meta}{template metadata object}

  \item{include.examples}{include examples in results}
}
\description{
  Converts template metadata to character vector.
}

