\name{check.input.class}
\alias{check.input.class}
\title{Check input value class}
\usage{
  check.input.class(value,
    class = c("character", "complex", "factor", "integer", "logical", "numeric", "raw", "option", "any"),
    input.name = NULL)
}
\arguments{
  \item{input}{}

  \item{class}{}
}
\description{
  Checks the class of value
}

