% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputs.R
\name{check.input.value.class}
\alias{check.input.value.class}
\title{Check Input Value Class}
\usage{
check.input.value.class(value, class = c("character", "complex", "factor",
  "integer", "logical", "numeric", "raw"), input.name = NULL)
}
\arguments{
\item{value}{input value}

\item{class}{input class (defaults to \code{NULL})}

\item{input.name}{input name (used in messages)}
}
\description{
Checks the class of an input value.
}

