\name{check.input.value}
\alias{check.input.value}
\title{Check input value}
\usage{
  check.input.value(input, value = NULL,
    attribute.name = c("length", "nchar", "nlevels", "limit"))
}
\arguments{
  \item{input}{}

  \item{value}{}

  \item{attribute.name}{}
}
\description{
  Check input value
}

