\name{check.report.chunks}
\alias{check.report.chunks}
\title{Check Report Chunks}
\usage{
  check.report.chunks(rp,
    what = c("errors", "warnings", "messages"))
}
\arguments{
  \item{rp}{\code{rapport} object}

  \item{what}{what fields to check. defaults to all}
}
\description{
  Checks for warnings and errors in report chunks.
}

