\name{grid.y}
\alias{grid.y}
\title{Add grid to y axis}
\usage{
  grid.y(side = c("top", "bottom", "left", "right"), ...,
    ticks = c("default", "yes", "no"), scales, components,
    line.col)
}
\arguments{
  \item{side}{}

  \item{...}{}

  \item{ticks}{}

  \item{scales}{}

  \item{components}{}

  \item{line.col}{}
}
\description{
  Add grid to y axis
}
\references{
  Forked from \code{latticeExtra::axis.grid()}.
}
\keyword{internal}

