\name{guess.input.description}
\alias{guess.input.description}
\title{Input Description}
\usage{
  guess.input.description(description, ...)
}
\arguments{
  \item{description}{a character string containing input
  description}

  \item{...}{additional arguments to be passed to
  \code{\link{grepl}}}
}
\description{
  Checks input description.
}

