% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputs.R
\name{guess.input}
\alias{guess.input}
\title{Guess Input}
\usage{
guess.input(input)
}
\arguments{
\item{input}{a named list containing input definition}
}
\description{
Checks and returns valid input from YAML input definition.
}

