\name{guess.old.input.length}
\alias{guess.old.input.length}
\title{Deprecated input limits}
\usage{
  guess.old.input.length(x, input.type)
}
\arguments{
  \item{x}{a character string containing input length
  definition}

  \item{input.type}{a character string containing input
  type}
}
\description{
  Guess deprecated input length.
}

