% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{guess.old.input.type}
\alias{guess.old.input.type}
\title{Check Type}
\usage{
guess.old.input.type(x)
}
\arguments{
\item{x}{a character string containing input definition}
}
\description{
Checks type of template input, based on provided sting. If input definition is syntactically correct, a list is returned, containing input type, size limits, and default value (for CSV options and boolean types only).
}

