\name{has.tags}
\alias{has.tags}
\title{Tag Existence}
\usage{
  has.tags(x, ...)
}
\arguments{
  \item{x}{a character value to check for \code{brew}
  strings}

  \item{...}{additional arguments for \code{\link{grepl}}
  function}
}
\value{
  a logical value indicating if the string has passed the
  check
}
\description{
  Checks if a character value contains specified tags.
}

