\name{htest}
\alias{htest}
\title{Hypothesis Test Helper}
\usage{
  htest(x, test, ...)
}
\arguments{
  \item{x}{arguments to be passed to function specified in
  \code{test}}

  \item{test}{a function to be applied}

  \item{...}{additional arguments for function specified in
  \code{test}}
}
\description{
  This function extracts only the most important
  information from \code{htest} class objects.
}

