\name{is.heading}
\alias{is.heading}
\title{Pandoc Heading}
\usage{
  is.heading(x)
}
\arguments{
  \item{x}{a string to test for pandoc heading format}
}
\value{
  a logical value indicating the string is (not) a pandoc
  heading
}
\description{
  Checks if provided string is a valid ATX-style pandoc
  heading.
}

