\name{is.rp.block}
\alias{is.rp.block}
\title{Rapport Block Element}
\usage{
  is.rp.block(x)
}
\arguments{
  \item{x}{any R object to check}
}
\value{
  a logical value indicating whether provided object is a
  \code{rp.block} object
}
\description{
  Checks if provided R object is a \code{rapport} block
  element.
}

