\name{is.rp.heading}
\alias{is.rp.heading}
\title{Rapport Heading Element}
\usage{
  is.rp.heading(x)
}
\arguments{
  \item{x}{any R object to check}
}
\value{
  a logical value indicating whether provided object is a
  \code{rp.heading} object
}
\description{
  Checks if provided R object is a \code{rapport} inline
  element.
}

