\name{is.rp.inline}
\alias{is.rp.inline}
\title{Rapport Inline Element}
\usage{
  is.rp.inline(x)
}
\arguments{
  \item{x}{any R object to check}
}
\value{
  a logical value indicating whether provided object is a
  \code{rp.inline} object
}
\description{
  Checks if provided R object is a \code{rapport} inline
  element.
}

