\name{is.tabular}
\alias{is.tabular}
\title{Tabular Structure}
\usage{
  is.tabular(x)
}
\arguments{
  \item{x}{an object to be checked for "tabular" format}
}
\value{
  a logical value indicating that provided object has
  tabular structure
}
\description{
  Checks if object has "tabular" structure - in this
  particular case, that means \code{\link{matrix}} and
  \code{\link{data.frame}} classes only.
}

