\docType{data}
\name{ius2008}
\alias{ius2008}
\title{Internet Usage Survey}
\description{
  This dataset contains data gathered in a survey of
  Internet usage in Serbian population in the period from
  April to May 2008.
}
\examples{
ius2008
}
\author{
  Aleksandar Blagotić

  Dušan Vučković
}
\keyword{datasets}

