\name{pandoc.emphasis}
\alias{pandoc.emphasis}
\title{Add emphasis}
\usage{
  pandoc.emphasis(x)
}
\arguments{
  \item{x}{character vector}
}
\value{
  character vector
}
\description{
  Pandoc style emphasis format (e.g. \code{*FOO*}) is added
  to character string.
}
\references{
  John MacFarlane (2012): _Pandoc User's Guide_.
  \url{http://johnmacfarlane.net/pandoc/README.html}
}
\seealso{
  \code{\link{pandoc.strong}}
  \code{\link{pandoc.strikeout}}
  \code{\link{pandoc.verbatim}}
}

