\name{pandoc.horizontal.rule}
\alias{pandoc.horizontal.rule}
\title{Create horizontal rule}
\usage{
  pandoc.horizontal.rule()
}
\value{
  character vector
}
\description{
  Create horizontal rule
}
\references{
  John MacFarlane (2012): _Pandoc User's Guide_.
  \url{http://johnmacfarlane.net/pandoc/README.html}
}

