\name{pandoc.strikeout}
\alias{pandoc.strikeout}
\title{Add strikeout}
\usage{
  pandoc.strikeout(x)
}
\arguments{
  \item{x}{character vector}
}
\value{
  character vector
}
\description{
  Pandoc style strikeout format (e.g. \code{~~FOO~~}) is
  added to character string.
}
\references{
  John MacFarlane (2012): _Pandoc User's Guide_.
  \url{http://johnmacfarlane.net/pandoc/README.html}
}
\seealso{
  \code{\link{pandoc.emphasis}} \code{\link{pandoc.strong}}
  \code{\link{pandoc.verbatim}}
}

