\name{pandoc.verbatim}
\alias{pandoc.verbatim}
\title{Add verbatim}
\usage{
  pandoc.verbatim(x)
}
\arguments{
  \item{x}{character vector}
}
\value{
  character vector
}
\description{
  Pandoc style verbatim format (e.g. \code{`FOO`}) is added
  to character string.
}
\references{
  John MacFarlane (2012): _Pandoc User's Guide_.
  \url{http://johnmacfarlane.net/pandoc/README.html}
}
\seealso{
  \code{\link{pandoc.emphasis}}
  \code{\link{pandoc.strikeout}}
  \code{\link{pandoc.strong}}
}

