\name{print.rapport.info}
\alias{print.rapport.info}
\title{Print Template Header}
\usage{
\method{print}{rapport.info}(x, ...)
}
\arguments{
  \item{x}{object of class \code{rp.header}. See
  \code{\link{rapport.header}} for details.}

  \item{...}{ignored}
}
\description{
Prints out the contents of template header (both metadata
and inputs) in human-readable format, so you can get
insight about the template requirements.
}

