% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.methods.R
\name{print.rapport.inputs}
\alias{print.rapport.inputs}
\title{Print Template Inputs}
\usage{
\method{print}{rapport.inputs}(x, ...)
}
\arguments{
\item{x}{object of class \code{rapport.inputs}. See \code{\link{rapport.inputs}} for details.}

\item{...}{ignored}
}
\description{
Prints out the contents of template inputs in human-readable format.
}
