% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.methods.R
\name{print.rapport.meta}
\alias{print.rapport.meta}
\title{Print Template Metadata}
\usage{
\method{print}{rapport.meta}(x, ...)
}
\arguments{
\item{x}{object of class \code{rapport.meta}. See \code{\link{rapport.meta}} for details.}

\item{...}{ignored}
}
\description{
Prints out the contents of template metadata in human-readable format.
}
