\name{print.rappot.meta}
\alias{print.rappot.meta}
\title{Print Template Metadata}
\usage{
\method{print}{rapport.meta}(x, ...)
}
\arguments{
  \item{x}{object of class \code{rapport.meta}. See
  \code{\link{rapport.meta}} for details.}

  \item{...}{ignored}
}
\description{
Prints out the contents of template metadata in
human-readable format.
}

