\name{print.rp.inputs}
\alias{print.rp.inputs}
\title{Print Template Inputs}
\usage{
  print.rp.inputs(x, type = c("text", "pandoc"))
}
\arguments{
  \item{x}{object of class \code{rp.inputs}. See
  \code{\link{tpl.inputs}} for details.}

  \item{type}{a string with output format. Defaults to
  plain text output.}
}
\description{
  Prints out the contents of template inputs in
  human-readable format.
}

