\name{print.rp.meta}
\alias{print.rp.meta}
\title{Print Template Metadata}
\usage{
  \method{print}{rp.meta} (x, type = c("text", "pandoc"))
}
\arguments{
  \item{x}{object of class \code{rp.meta}. See
  \code{\link{tpl.meta}} for details.}

  \item{type}{a string with output format. Defaults to
  plain text output.}
}
\description{
  Prints out the contents of template metadata in
  human-readable format.
}

