% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/template.R
\name{rapport.info}
\alias{rapport.info}
\alias{tpl.info}
\title{Template Info}
\usage{
rapport.info(fp, meta = TRUE, inputs = TRUE)
}
\arguments{
\item{fp}{a template file pointer (see \code{rapport:::rapport.read} for details)}

\item{meta}{return template metadata? (defaults to \code{TRUE})}

\item{inputs}{return template inputs? (defaults to \code{TRUE})}
}
\description{
Provides information about template metadata and/or inputs. See \code{\link{rapport.meta}} and \code{\link{rapport.inputs}} for details.
}
\examples{
\dontrun{
rapport.info('Example')                    # return both metadata and inputs
rapport.info('Crosstable', inputs = FALSE) # return only template metadata
rapport.info('Correlation', meta = FALSE)  # return only template inputs
}
}
\seealso{
{
\code{\link{rapport.meta}}
\code{\link{rapport.inputs}}
}
}

