% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{rapport.odt}
\alias{rapport.odt}
\title{Rapport to ODT}
\usage{
rapport.odt(...)
}
\arguments{
\item{...}{parameters passed directly to \code{\link{rapport}}}
}
\description{
This is a simple wrapper around \code{\link{rapport}} and \code{\link{rapport.export}}. Basically it works like \code{\link{rapport}} but the returned class is exported at one go.
}
\seealso{
\code{\link{rapport.export}} \code{\link{rapport.html}} \code{\link{rapport.pdf}} \code{\link{rapport.docx}}
}
