% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rp_helpers.R
\name{rapport.path}
\alias{rapport.path}
\alias{tpl.paths}
\title{Template Paths}
\usage{
rapport.path()
}
\value{
a character vector with paths
}
\description{
List all custom paths where rapport will look for templates.
}
\examples{
\dontrun{
rapport.path()
}
}

