% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_helpers.R
\name{rapport.path.reset}
\alias{rapport.path.reset}
\alias{tpl.paths.reset}
\title{Reset Template Paths}
\usage{
rapport.path.reset()
}
\description{
Resets to default (NULL) all custom paths where rapport will look for templates.
}
\examples{
\dontrun{
rapport.path.reset()
}
}
