\name{rapport.pdf}
\alias{rapport.pdf}
\title{Rapport to PDF}
\usage{
  rapport.pdf(...)
}
\arguments{
  \item{...}{parameters passed directly to
  \code{\link{rapport}}}
}
\description{
  This is a simple wrapper around \code{\link{rapport}} and
  \code{\link{tpl.export}}. Basically it works like
  \code{\link{rapport}} but the returned class is exported
  at one go.
}

