% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{rapport.tangle}
\alias{rapport.tangle}
\alias{tpl.tangle}
\title{Extract template chunk contents}
\usage{
rapport.tangle(fp, file = "", show.inline.chunks = FALSE)
}
\arguments{
\item{fp}{template file pointer (see \code{rapport:::rapport.read} for details)}

\item{file}{see \code{file} argument in \code{\link{cat}} function documentation}

\item{show.inline.chunks}{extract contents of inline chunks as well? (defaults to \code{FALSE})}
}
\value{
(invisibly) a list with either inline or block chunk contents
}
\description{
\code{rapport}'s alternative to \code{\link{Stangle}} - extracts contents of template chunks. If \code{file} argument
}
