\name{report.matrix}
\alias{report.matrix}
\title{Report a matrix}
\usage{
  report.matrix(x, tpl = "classes/matrix.tpl", ...)
}
\arguments{
  \item{x}{}

  \item{tpl}{}

  \item{...}{}
}
\description{
  Report a matrix
}
\examples{
\dontrun{
report(matrix(1:25, 5, 5))
report(matrix(1:25/3, 5, 5))
}
}

