\name{rp.export}
\alias{rp.export}
\title{Export rapport class (deprecated)}
\usage{
  rp.export(...)
}
\arguments{
  \item{...}{parameters passed to \code{\link{tpl.export}}}
}
\description{
  This function points to \code{\link{tpl.export}}. Use
  that insted \code{rp.export}, as this function remained
  here only for backward compatibily of the package.
}

