\name{rp.graph.check}
\alias{rp.graph.check}
\title{Input cheks (internal)}
\usage{
  rp.graph.check(x, facet = NULL, subset = NULL, ...)
}
\arguments{
  \item{x}{a variable}

  \item{facet}{if facet set}

  \item{subset}{if subset set}

  \item{...}{other parameters}
}
\description{
  Internal function used by eg. \code{rp.histogram}.
}
\keyword{internal}

