\name{rp.iqr}
\alias{rp.iqr}
\title{Interquartile Range}
\usage{
  rp.iqr(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with variable's interquartile range
}
\description{
  Calculates interquartile range of given variable. See
  \code{\link{rp.univar}} for details.
}

