\name{rp.kurtosis}
\alias{rp.kurtosis}
\title{Kurtosis}
\usage{
  rp.kurtosis(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with variable's kurtosis
}
\description{
  Calculates kurtosis of given variable. See
  \code{\link{rp.univar}} for details.
}

