\name{rp.max}
\alias{rp.max}
\title{Maximum}
\usage{
  rp.max(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with maximum value
}
\description{
  Returns the maximum of all values in a vector by passing
  \{code{\link{max}} as \code{fn} argument to
  \code{\link{rp.univar}} function.
}

