\name{rp.mean}
\alias{rp.mean}
\title{Mean}
\usage{
  rp.mean(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with variable's mean
}
\description{
  Calculates mean of given variable by passing
  \code{\link{sum}} as \code{fn} argument to
  \code{\link{rp.univar}} function.
}

