\name{rp.median}
\alias{rp.median}
\title{Median}
\usage{
  rp.median(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with variable's median
}
\description{
  Calculates median of given variable. See
  \code{\link{rp.univar}} for details.
}

