\name{rp.min}
\alias{rp.min}
\title{Minimum}
\usage{
  rp.min(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with minimum value
}
\description{
  Returns the minimum of all values in a vector by passing
  \{code{\link{min}} as \code{fn} argument to
  \code{\link{rp.univar}} function.
}

