\name{rp.range}
\alias{rp.range}
\title{Range}
\usage{
  rp.range(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with calculated range
}
\description{
  Calculates difference between the largest and the
  smallest value in a vector. See \code{\link{rp.univar}}
  for details.
}

