\name{rp.sd}
\alias{rp.sd}
\title{Standard Deviation}
\usage{
  rp.sd(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with variable's standard deviation
}
\description{
  Calculates standard deviation of given variable. See
  \code{\link{rp.univar}} for details.
}

