\name{rp.se.mean}
\alias{rp.se.mean}
\title{Standard Error of Mean}
\usage{
  rp.se.mean(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with standard error of mean
}
\description{
  Calculates standard error of mean for given variable. See
  \code{\link{rp.univar}} for details.
}

