\name{rp.skewness}
\alias{rp.skewness}
\title{Skewness}
\usage{
  rp.skewness(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with variable's skewness
}
\description{
  Calculates skewness of given variable. See
  \code{\link{rp.univar}} for details.
}

