\name{rp.sum}
\alias{rp.sum}
\title{Sum}
\usage{
  rp.sum(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with sum of vector elements
}
\description{
  Returns the sum of variable's elements, by passing
  \code{\link{sum}} as \code{fn} argument to
  \code{\link{rp.univar}} function.
}

