\name{rp.var}
\alias{rp.var}
\title{Variance}
\usage{
  rp.var(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{rp.univar}
  function}
}
\value{
  a numeric value with variable's variance
}
\description{
  Calculates variance of given variable. See
  \code{\link{rp.univar}} for details.
}

