% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rp_helpers.R
\name{as.character.rapport.inputs}
\alias{as.character.rapport.inputs}
\title{Convert Inputs to Character}
\usage{
\method{as.character}{rapport.inputs}(x, ...)
}
\arguments{
\item{x}{template inputs object}

\item{...}{ignored}
}
\description{
Converts template inputs to character vector with YAML strings.
}

