\name{as.character.rp.meta}
\alias{as.character.rp.meta}
\title{Convert Metadata to Character}
\usage{
  \method{as.character}{rp.meta} (x, ...)
}
\arguments{
  \item{x}{template metadata object}

  \item{...}{ignored}
}
\description{
  Converts template metadata to character vector with YAML
  strings.
}

