\name{as.character.rp.meta}
\alias{as.character.rp.meta}
\title{Convert Metadata to Character}
\usage{
  \method{as.character}{rp.meta} (x, ...)
}
\arguments{
  \item{x}{template metadata object}

  \item{...}{accepts \code{include.examples} which
  indicates that examples should be included in output (if
  any)}
}
\description{
  Converts template metadata to character vector.
}

