% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputs.R
\name{guess.input.description}
\alias{guess.input.description}
\title{Input Description}
\usage{
guess.input.description(description)
}
\arguments{
\item{description}{a character string containing input description}
}
\description{
Checks and returns input description.
}

