% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{guess.input.label}
\alias{guess.input.label}
\title{Input Label}
\usage{
guess.input.label(label)
}
\arguments{
\item{label}{a character string containing input label}
}
\description{
Checks and returns input label.
}
