\name{htest}
\alias{htest}
\title{Hypothesis Tests}
\usage{
  htest(x, ...)
}
\arguments{
  \item{x}{arguments to be passed to function specified in
  \code{test}}

  \item{...}{additional arguments for function specified in
  \code{test}}
}
\description{
  This function extracts only the most important
  information from \code{htest} class objects - statistic
  and its p-value.
}

