% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rp_helpers.R
\name{is.rapport}
\alias{is.rapport}
\title{Rapport Object}
\usage{
is.rapport(x)
}
\arguments{
\item{x}{any R object to check}
}
\value{
a logical value indicating whether provided object is a \code{rapport} object
}
\description{
Checks if provided R object is of \code{rapport} class.
}

