\name{print.rp.info}
\alias{print.rp.info}
\title{Print Template Header}
\usage{
  \method{print}{rp.info} (x, ...)
}
\arguments{
  \item{x}{object of class \code{rp.header}. See
  \code{\link{tpl.header}} for details.}

  \item{...}{ignored}
}
\description{
  Prints out the contents of template header (both metadata
  and inputs) in human-readable format, so you can get
  insight about the template requirements.
}

