\name{print.rp.inputs}
\alias{print.rp.inputs}
\title{Print Template Inputs}
\usage{
  \method{print}{rp.inputs} (x, ...)
}
\arguments{
  \item{x}{object of class \code{rp.inputs}. See
  \code{\link{tpl.inputs}} for details.}

  \item{...}{ignored}
}
\description{
  Prints out the contents of template inputs in
  human-readable format.
}

