\name{print.rp.meta}
\alias{print.rp.meta}
\title{Print Template Metadata}
\usage{
  \method{print}{rp.meta} (x, ...)
}
\arguments{
  \item{x}{object of class \code{rp.meta}. See
  \code{\link{tpl.meta}} for details.}

  \item{...}{ignored}
}
\description{
  Prints out the contents of template metadata in
  human-readable format.
}

