% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport.R
\name{rapport-helpers}
\alias{rapport-helpers}
\title{\emph{rapport helpers}}
\description{
\code{rapport} package comes with bunch of helper functions that make your template writing and report creation easier, although most of these helpers were migrated to the \code{rapportools} package.
}
\details{
\emph{Export helpers}

\itemize{
    \item \code{\link{rapport.docx}}
    \item \code{\link{rapport.html}}
    \item \code{\link{rapport.odt}}
    \item \code{\link{rapport.pdf}}
    \item \code{\link{rapport.export}}
}

Please load the \code{rapportools} package if you would use any of the below functions in the \code{.GlobalEnv}, or simply add \code{rapportools} to the required packages section in your template file. That latter is a lot cleaner solution.

\emph{General purpose helpers}

\itemize{
    \item \code{\link{adj.rle}}
    \item \code{\link{alike.integer}}
    \item \code{\link{capitalise}}
    \item \code{\link{catn}}
    \item \code{\link{fml}}
    \item \code{\link{is.boolean}}
    \item \code{\link{is.empty}}
    \item \code{\link{is.number}}
    \item \code{\link{is.string}}
    \item \code{\link{is.tabular}}
    \item \code{\link{is.variable}}
    \item \code{\link{messagef}}
    \item \code{\link{pct}}
    \item \code{\link{stopf}}
    \item \code{\link{tocamel}}
    \item \code{\link{trim.space}}
    \item \code{\link{vgsub}}
    \item \code{\link{warningf}}
}

\emph{Summary statistics}

\itemize{
    \item \code{\link{rp.desc}}
    \item \code{\link{rp.freq}}
}

\emph{Univariate descriptive statistics}

\itemize{
    \item \code{\link{rp.iqr}}
    \item \code{\link{rp.max}}
    \item \code{\link{rp.mean}}
    \item \code{\link{rp.median}}
    \item \code{\link{rp.min}}
    \item \code{\link{rp.missing}}
    \item \code{\link{rp.percent}}
    \item \code{\link{rp.range}}
    \item \code{\link{rp.sd}}
    \item \code{\link{rp.se.mean}}
    \item \code{\link{rp.sum}}
    \item \code{\link{rp.valid}}
    \item \code{\link{rp.var}}
}

\emph{Miscelaneous stats helpers}

\itemize{
    \item \code{\link{htest}}
    \item \code{\link{htest.short}}
    \item \code{\link{kurtosis}}
    \item \code{\link{skewness}}
    \item \code{\link{lambda.test}}
    \item \code{\link{rp.outlier}}
}
}
